Trait for BaseController
