+++ app/Service/BaseService.php
@@ -7,6 +7,8 @@ use Illuminate\Foundation\Http\FormRequest;
 
 abstract class BaseService
 {
+    use CrudOperations;
+
     protected RepositoryInterface $eloquent;
 
     protected ModelInterface $model;
@@ -17,39 +19,8 @@ abstract class BaseService
         $this->model = $model;
     }
 
-    public function index(): string
+    protected function getEloquent(): RepositoryInterface
     {
-        return $this
-            ->eloquent
-            ->getAll()
-            ->toJson();
-    }
-
-    public function show(ModelInterface $model): string
-    {
-        return $model->toJson();
-    }
-
-    public function store(FormRequest $request): string
-    {
-        return $this
-            ->eloquent
-            ->store($request->validated())
-            ->toJson();
-    }
-
-    public function update(FormRequest $request, ModelInterface $model): string
-    {
-        return $this
-            ->eloquent
-            ->update($request->validated(), $model)
-            ->toJson();
-    }
-
-    public function destroy(ModelInterface $model): bool
-    {
-        return $this
-            ->eloquent
-            ->destroy($model);
+        return $this->eloquent;
     }
 }
