+++ routes/web.php
@@ -18,6 +18,7 @@ use Illuminate\Support\Facades\Route;
 Route::group(['middleware' => 'api', 'prefix' => 'laravel_api'], function () {
     Route::resource('questions', 'App\Http\Controllers\QuestionController');
     Route::resource('answers', 'App\Http\Controllers\AnswerController');
+//    Route::get('questions/{id}/answers', 'App\Http\Controllers\AnswerController@index');
 });
